patches by [C]167:
[PATCH 1/2] Fix missing include causing errors under g++4.3

Using libstdc++ 4.2, no include to <algorithm> is needed in order to
have std::find.
This seems to have changed in libstdc++ 4.3 (Debian SID at least), so
the header has to be included explicitly.
The code now compiles under versions 4.1, 4.2 and 4.3 of the GNU
Compiler Collection.

Subject: [PATCH 2/2] Fix warnings when compiled with g++ 4.3

git-svn-id: http://psotnic.svn.sourceforge.net/svnroot/psotnic@196 642afdc4-8138-0410-ae4e-e573936c0dc0